requirements.txt
aiosmtplib==2.0.2
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
bcrypt
click==8.1.7
dnspython==2.6.1
dotenv
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.112.2
fastapi-mail==1.4.1
greenlet
h11==0.14.0
idna==3.8
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2-binary
pyasn1==0.6.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
typing_extensions==4.12.2
uuid7==0.1.0
uvicorn==0.30.6
